requirements.txt
apache-airflow==3.2.0
beautifulsoup4==4.13.4
lxml==5.4.0
python-dateutil==2.9.0.post0
requests==2.32.5
pydantic==2.12.5
pydantic-settings==2.13.1
SQLAlchemy==2.0.49
psycopg2-binary==2.9.11
loguru==0.7.3
markdownify==1.1.0
psutil==7.2.2
PyYAML==6.0.3
prefect==3.6.27
langchain-text-splitters==1.1.2
qdrant-client==1.17.1
fastembed==0.8.0
huggingface-hub==1.11.0

opik==2.0.2
gradio==6.12.0
